requirements.txt
fastapi
uvicorn
aiogram
python-dotenv
sqlalchemy
asyncpg
razorpay
httpx
python-dateutil

